ft.otp/requirements.txt
Flask==2.2.3
debugpy==1.6.7
bootstrap-flask==2.2.0
Flask-FontAwesome==0.1.5
pycryptodome ==3.17
pytest==7.3.1
mypy==1.2.0
black==23.3.0
ruff==0.0.261